# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Radio call manager API — real-time audio streaming, call management, and admin CRUD.",
        "title": "OpenScanner API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/bookmarks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns an array of call IDs that the authenticated user has bookmarked.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmarks"
                ],
                "summary": "List bookmarked call IDs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.BookmarkIDsResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a bookmark if one does not exist for the given call and user, or removes it if it already exists.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Toggle bookmark on a call",
                "parameters": [
                    {
                        "description": "Call to bookmark",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.ToggleBookmarkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ToggleBookmarkResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/bookmarks/calls": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns full call details for all calls bookmarked by the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmarks"
                ],
                "summary": "List bookmarked calls with metadata",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.BookmarkCallsResponse"
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.BookmarkCallsResponse": {
            "type": "object",
            "properties": {
                "calls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.CallSearchResult"
                    }
                }
            }
        },
        "internal_api.BookmarkIDsResponse": {
            "type": "object",
            "properties": {
                "callIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "internal_api.CallSearchResult": {
            "type": "object",
            "properties": {
                "audioName": {
                    "type": "string"
                },
                "audioType": {
                    "type": "string"
                },
                "bookmarked": {
                    "type": "boolean"
                },
                "channel": {
                    "type": "string"
                },
                "dateTime": {
                    "type": "integer"
                },
                "decoder": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "errorCount": {
                    "type": "integer"
                },
                "frequency": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "site": {
                    "type": "string"
                },
                "source": {
                    "type": "integer"
                },
                "spikeCount": {
                    "type": "integer"
                },
                "systemId": {
                    "type": "integer"
                },
                "systemLabel": {
                    "type": "string"
                },
                "talkgroupGroup": {
                    "type": "string"
                },
                "talkgroupId": {
                    "type": "integer"
                },
                "talkgroupLabel": {
                    "type": "string"
                },
                "talkgroupLed": {
                    "type": "string"
                },
                "talkgroupName": {
                    "type": "string"
                },
                "talkgroupTag": {
                    "type": "string"
                },
                "transcript": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "descriptive error message"
                }
            }
        },
...
        "internal_api.ToggleBookmarkRequest": {
            "type": "object",
            "properties": {
                "callId": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
        "internal_api.ToggleBookmarkResponse": {
            "type": "object",
            "properties": {
                "bookmarked": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "integer",
                    "example": 7
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "API key for call upload endpoints",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT token prefixed with \"Bearer \"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}